+++ packages/babel-preset-gatsby/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.9.0](https://github.com/gatsbyjs/gatsby/commits/babel-preset-gatsby@3.9.0/packages/babel-preset-gatsby) (2023-04-18)
+
+[🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v5.9)
+
+**Note:** Version bump only for package babel-preset-gatsby
+
 ## [3.8.0](https://github.com/gatsbyjs/gatsby/commits/babel-preset-gatsby@3.8.0/packages/babel-preset-gatsby) (2023-03-21)
 
 [🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v5.8)