+++ packages/gatsby-plugin-preact/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.9.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-plugin-preact@7.9.0/packages/gatsby-plugin-preact) (2023-04-18)
+
+[🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v5.9)
+
+#### Chores
+
+- update dependency preact to ^10.13.2 for gatsby-plugin-preact [#37844](https://github.com/gatsbyjs/gatsby/issues/37844) ([78804bd](https://github.com/gatsbyjs/gatsby/commit/78804bd357b3cb51c01a31ce79dcde43873e56dc))
+
 ## [7.8.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-plugin-preact@7.8.0/packages/gatsby-plugin-preact) (2023-03-21)
 
 [🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v5.8)