+++ packages/gatsby-source-shopify/CHANGELOG.md
@@ -3,6 +3,27 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [8.9.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-source-shopify@8.9.0/packages/gatsby-source-shopify) (2023-04-18)
+
+[🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v5.9)
+
+#### Features
+
+- use gatsby node id to enable relationships [#37775](https://github.com/gatsbyjs/gatsby/issues/37775) ([a2c3fa3](https://github.com/gatsbyjs/gatsby/commit/a2c3fa3f8af6ec7021bbdcb207d886ea16304dfd))
+
+#### Bug Fixes
+
+- update dependency prettier to v2.8.7 [#37834](https://github.com/gatsbyjs/gatsby/issues/37834) ([029ee85](https://github.com/gatsbyjs/gatsby/commit/029ee85e3ea3dd819f07159fcb22b4fb35ebbf86))
+- update dependency node-fetch to ^2.6.9 for gatsby-source-shopify [#37830](https://github.com/gatsbyjs/gatsby/issues/37830) ([ae4a765](https://github.com/gatsbyjs/gatsby/commit/ae4a765ff02c86fddcb9ae655699754e40725752))
+- update typescript to v5 (major) [#37906](https://github.com/gatsbyjs/gatsby/issues/37906) ([3859ba6](https://github.com/gatsbyjs/gatsby/commit/3859ba6ce629e658525f858a8799b08c34c3eda5))
+- Correct interface query [#37788](https://github.com/gatsbyjs/gatsby/issues/37788) ([ae1e8e9](https://github.com/gatsbyjs/gatsby/commit/ae1e8e98f323a9ee04a71a46eddf453ff6cd6ccd))
+
+#### Chores
+
+- update [dev] major dependencies for gatsby-source-shopify (major) [#37889](https://github.com/gatsbyjs/gatsby/issues/37889) ([1fa8314](https://github.com/gatsbyjs/gatsby/commit/1fa831481cfd02c7d50377b536d9ca516c6c6df9))
+- update typescript [#37848](https://github.com/gatsbyjs/gatsby/issues/37848) ([9f92796](https://github.com/gatsbyjs/gatsby/commit/9f92796070b0e73fa92089af93de376c9be9021e))
+- update [dev] minor and patch dependencies for gatsby-source-shopify [#37818](https://github.com/gatsbyjs/gatsby/issues/37818) ([4def2f4](https://github.com/gatsbyjs/gatsby/commit/4def2f4f92c97dc99d5f5b9b823710318b529c33))
+
 ### [8.8.2](https://github.com/gatsbyjs/gatsby/commits/gatsby-source-shopify@8.8.2/packages/gatsby-source-shopify) (2023-03-28)
 
 #### Bug Fixes