+++ packages/gatsby-telemetry/CHANGELOG.md
@@ -3,6 +3,21 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.9.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-telemetry@4.9.0/packages/gatsby-telemetry) (2023-04-18)
+
+[🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v5.9)
+
+#### Bug Fixes
+
+- update dependency node-fetch to ^2.6.9 for gatsby-telemetry [#37833](https://github.com/gatsbyjs/gatsby/issues/37833) ([5f04821](https://github.com/gatsbyjs/gatsby/commit/5f0482164e3aea7de0a9a0c40f44555d16402e9f))
+- update dependency fs-extra to ^11.1.1 [#37827](https://github.com/gatsbyjs/gatsby/issues/37827) ([3e9a590](https://github.com/gatsbyjs/gatsby/commit/3e9a590fdc0b4cdbc763b6a6e3ca3e983ac8231a))
+- update typescript to v5 (major) [#37906](https://github.com/gatsbyjs/gatsby/issues/37906) ([3859ba6](https://github.com/gatsbyjs/gatsby/commit/3859ba6ce629e658525f858a8799b08c34c3eda5))
+
+#### Chores
+
+- update typescript [#37848](https://github.com/gatsbyjs/gatsby/issues/37848) ([9f92796](https://github.com/gatsbyjs/gatsby/commit/9f92796070b0e73fa92089af93de376c9be9021e))
+- update dependency rimraf to v4 for gatsby-telemetry [#37900](https://github.com/gatsbyjs/gatsby/issues/37900) ([617a3d4](https://github.com/gatsbyjs/gatsby/commit/617a3d4e87c1ec061bc877173b158cdcb64f367d))
+
 ## [4.8.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-telemetry@4.8.0/packages/gatsby-telemetry) (2023-03-21)
 
 [🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v5.8)